Move buildQueryString file to root
Also split the file into two.
