+++ src/index.ts
@@ -1,13 +1,15 @@
-export { VALIDATORS } from './validators';
-
+/** Hooks */
 export { createUseQueryParamsStateHook } from './createUseQueryParamsStateHook';
-
 export { useQueryParam } from './useQueryParam';
 export { useQueryParamsState } from './useQueryParamsState';
+export { useBuildQueryStringFromCurrentURL } from './useBuildQueryStringFromCurrentURL';
 
-export * from './qparams';
-export * from './helpers';
-export * from './useBuildQueryStringFromCurrentURL';
+/** Helpers */
+export { QPARAMS } from './qparams';
+export { buildQueryString } from './buildQueryString';
+export { pickQueryParamsMatchingSchema } from './pickQueryParamsMatchingSchema';
 
+/** Other */
 export { QS_BUILD_STRATEGY } from './types';
 export { Errors } from './errors';
+export { VALIDATORS } from './validators';
