+++ src/internal/buildQueryStringFromCurrentURL.ts
@@ -5,13 +5,15 @@ import { IQueryParamsSchema, QueryParams, QS_BUILD_STRATEGY } from '../types';
 import {
   getAllRawQueryParamsFromURL,
   getExternalQueryParamsFromURL,
-} from '../internal/getQueryParamsFromURL';
+} from './getQueryParamsFromURL';
 
-import { runParamsValidatorsPartial } from '../validators';
-import { serializeQueryParamsValues } from '../internal/serializer/serialize';
-import { getDefaultQueryParamsState } from '../internal/getDefaultQueryParamsState';
-import { createQueryString } from '../internal/queryString';
+import { serializeQueryParamsValues } from './serializer/serialize';
+import { getDefaultQueryParamsState } from './getDefaultQueryParamsState';
+import { buildQueryString } from '../buildQueryString';
 
+/**
+ * @internal
+ */
 export function buildQueryStringFromCurrentURL<
   QueryParamsSchema extends IQueryParamsSchema
 >(
@@ -34,44 +36,6 @@ export function buildQueryStringFromCurrentURL<
     contextData,
     rawQueryParamsMergeDestination
   );
-}
-
-/**
- * Create a new query string based on the provided schema and a matching query params object.
- * An error will be thrown if the param does not match the type defined in the schema.
- */
-export function buildQueryString<QueryParamsSchema extends IQueryParamsSchema>(
-  /**
-   * The parameters schema followed by the query string.
-   */
-  queryParamsSchema: QueryParamsSchema,
-  /**
-   * The query parameters you want to add to the query string and belong
-   * to the schema.
-   */
-  newQueryParams: Partial<QueryParams<QueryParamsSchema>> = {},
-  contextData?: any,
-  /**
-   * The query parameters you want to add to the query string but don't
-   * belong to the schema. Each parameter value is expected to already be in its
-   * string format.
-   */
-  otherRawParams?: Record<string, string | null | undefined>
-): string {
-  // Raise a JS error if we are trying to set a value that doesn't pass the validator
-  runParamsValidatorsPartial(
-    queryParamsSchema,
-    newQueryParams,
-    contextData,
-    /** throwOnError */ true
-  );
-
-  const serializedQueryParams = {
-    ...otherRawParams,
-    ...serializeQueryParamsValues(queryParamsSchema, newQueryParams),
-  };
-
-  return createQueryString(serializedQueryParams);
 }
 
 /**
