+++ src/useQueryParam.ts
@@ -21,6 +21,9 @@ export function useQueryParam<T>(
   contextData?: any
 ): [T | null | undefined, (value?: T | null | undefined) => void];
 
+/**
+ * Hook to manage one query param state at a time.
+ */
 export function useQueryParam(
   paramName: string,
   queryParamDef: QueryParamDef<any> = QPARAMS.string(),
